--- src-tauri/src/core/mod.rs
@@ -1,1 +1,0 @@
-pub mod setup;
