+++ src-tauri/src/lib.rs
@@ -1,20 +1,18 @@
 // Learn more about Tauri commands at https://tauri.app/develop/calling-rust/
-mod menu;
+mod transform;
 
 use anyhow::Result;
-use async_openai::{
-    config::OpenAIConfig,
-    types::CreateCompletionRequestArgs,
-    Client,
-};
 use dirs::config_dir;
 use serde::{Deserialize, Serialize};
 use std::{fs, path::PathBuf, sync::Mutex, collections::HashMap};
-use tauri::{Manager, menu::{Menu, MenuBuilder, MenuItemBuilder, MenuItem}, tray::{TrayIconBuilder}, Emitter};
+use tauri::{
+    Manager, 
+    menu::{MenuBuilder}, 
+    tray::{TrayIconBuilder},
+    Emitter
+};
 use tokio::sync::Mutex as TokioMutex;
 
-mod core;
-use crate::core::setup;
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
 pub struct Settings {
@@ -95,67 +93,6 @@ async fn get_settings(state: tauri::State<'_, AppState>) -> Result<Settings, Str
     Ok(state.settings.lock().await.clone())
 }
 
-async fn transform_text(text: &str, prompt: &str, api_key: &str) -> Result<String, String> {
-    println!("Starting text transformation with prompt: {}", prompt);
-
-    let config = OpenAIConfig::new().with_api_key(api_key);
-    let client = Client::with_config(config);
-    println!("OpenAI client created successfully");
-
-    let request = CreateCompletionRequestArgs::default()
-        .model("gpt-3.5-turbo-instruct")
-        .prompt(format!("{}\n\nText: {}", prompt, text))
-        .max_tokens(2000u16)
-        .temperature(0.7)
-        .build();
-
-    let request = match request {
-        Ok(req) => req,
-        Err(e) => {
-            let error = format!("Failed to build completion request: {}", e);
-            println!("{}", error);
-            return Err(error);
-        }
-    };
-
-    println!("Sending request to OpenAI...");
-    match client.completions().create(request).await {
-        Ok(response) => {
-            println!("Received response from OpenAI");
-            if let Some(choice) = response.choices.first() {
-                println!("Successfully transformed text");
-                Ok(choice.text.clone())
-            } else {
-                let error = "No completion choices returned from OpenAI".to_string();
-                println!("{}", error);
-                Err(error)
-            }
-        }
-        Err(e) => {
-            let error = format!("OpenAI API error: {}", e);
-            println!("{}", error);
-            Err(error)
-        }
-    }
-}
-
-#[tauri::command]
-async fn process_selected_text(
-    text: String,
-    state: tauri::State<'_, AppState>,
-) -> Result<String, String> {
-    let settings = state.settings.lock().await;
-
-    let prompt_key = settings.selected_tone.clone().unwrap_or_else(|| "Improve Writing".to_string());
-
-    let prompt = settings
-        .custom_prompts
-        .get(&prompt_key)
-        .ok_or_else(|| format!("Prompt '{}' not found", prompt_key))?;
-    let api_key = get_api_key().await?;
-    transform_text(&text, prompt, &api_key).await
-}
-
 #[tauri::command]
 async fn show_settings(window: tauri::Window) -> Result<(), String> {
     let app = window.app_handle();
@@ -170,38 +107,47 @@ async fn show_settings(window: tauri::Window) -> Result<(), String> {
 fn create_tray_menu(app: &tauri::App) -> Result<tauri::tray::TrayIcon, tauri::Error> {
     println!("Creating tray menu... 22");
 
-    let transform_i = MenuItem::with_id(app, "transform", "Transform", true, None::<&str>)?;
-    let settings_i = MenuItem::with_id(app, "settings", "Settings", true, None::<&str>)?;
-    let quit_i = MenuItem::with_id(app, "quit", "Quit", true, None::<&str>)?;
-
-    let menu = Menu::with_items(app, &[&transform_i, &settings_i, &quit_i])?;
+    println!("Creating menu...");
+    let menu = MenuBuilder::new(app)
+        .text("transform", "Transform")
+        .text("settings", "Settings")
+        .separator()
+        .text("quit", "Quit")
+        .build()?;
 
     let tray = TrayIconBuilder::new()
+        .icon(app.default_window_icon().unwrap().clone())
         .menu(&menu)
-        .menu_on_left_click(true)
-        .on_menu_event(|app, event| match event.id().as_ref() {
-            "quit" => {
-                println!("Quitting Milo app...");
-                app.exit(0);
-            }
-            "settings" => {
-                println!("Opening settings window...");
-                if let Some(window) = app.get_webview_window("main") {
-                    window.show().unwrap();
-                    window.set_focus().unwrap();
+        .show_menu_on_left_click(true)
+        .on_menu_event(|app, event| {
+            println!("Menu event received: {:?}", event.id());
+            match event.id().as_ref() {
+                "quit" => {
+                    println!("Quit menu item clicked");
+                    app.exit(0);
+                }
+                "settings" => {
+                    println!("Settings menu item clicked");
+                    if let Some(window) = app.get_webview_window("main") {
+                        window.show().unwrap();
+                        window.set_focus().unwrap();
+                    }
+                }
+                "transform" => {
+                    println!("Transform menu item clicked");
+                    let state = app.state::<AppState>();
+                    let is_transforming = *state.is_transforming.lock().unwrap();
+                    if !is_transforming {
+                        println!("Starting transformation...");
+                        app.emit("transform_clipboard", ()).unwrap();
+                    } else {
+                        println!("Transformation already in progress");
+                    }
+                }
+                _ => {
+                    println!("Unknown menu item clicked: {:?}", event.id());
                 }
             }
-            "transform" => {
-                let state = app.state::<AppState>();
-                let is_transforming = *state.is_transforming.lock().unwrap();
-                if !is_transforming {
-                    println!("Starting text transformation...");
-                    app.emit("transform_clipboard", ()).unwrap();
-                } else {
-                    println!("Text transformation already in progress...");
-                }
-            }
-            _ => {}
         })
         .build(app)?;
 
@@ -219,15 +165,7 @@ pub fn run() {
     tauri::Builder::default()
         .setup(|app| {
             println!("Starting Milo app...");
-            
-            // Get the main window and handle focus events
-            let main_window = app.get_webview_window("main").unwrap();
-            
-            // Create tray icon
-            create_tray_menu(app)?;
-
-            setup::default(app, main_window.clone());
-
+            let _tray = create_tray_menu(app)?;
             Ok(())
         })
         .manage(app_state)
@@ -236,9 +174,8 @@ pub fn run() {
             get_api_key,
             save_settings,
             get_settings,
-            process_selected_text,
             show_settings,
-            menu::transform_clipboard,
+            transform::transform_clipboard,
         ])
         .run(tauri::generate_context!())
         .expect("error while running tauri application");
