+++ src-tauri/src/transform.rs
@@ -1,17 +1,55 @@
 use arboard::Clipboard;
-use tauri::{Manager, menu::MenuBuilder, Emitter};
+use tauri::{Manager, Emitter};
+use async_openai::{
+    config::OpenAIConfig,
+    types::CreateCompletionRequestArgs,
+    Client,
+};
 
-pub fn create_app_menu(app: &tauri::App) -> Result<(), Box<dyn std::error::Error>> {
-    let menu = MenuBuilder::new(app)
-        .text("Transform", "transform")
-        .text("Settings", "settings")
-        .separator()
-        .text("Quit", "quit")
-        .build()?;
+async fn transform_text(text: &str, prompt: &str, api_key: &str) -> Result<String, String> {
+    println!("Starting text transformation with prompt: {}", prompt);
 
-    app.set_menu(menu)?;
-    Ok(())
+    let config = OpenAIConfig::new().with_api_key(api_key);
+    let client = Client::with_config(config);
+    println!("OpenAI client created successfully");
+
+    let request = CreateCompletionRequestArgs::default()
+        .model("gpt-3.5-turbo-instruct")
+        .prompt(format!("{}\n\nText: {}", prompt, text))
+        .max_tokens(2000u16)
+        .temperature(0.7)
+        .build();
+
+    let request = match request {
+        Ok(req) => req,
+        Err(e) => {
+            let error = format!("Failed to build completion request: {}", e);
+            println!("{}", error);
+            return Err(error);
+        }
+    };
+
+    println!("Sending request to OpenAI...");
+    match client.completions().create(request).await {
+        Ok(response) => {
+            println!("Received response from OpenAI");
+            if let Some(choice) = response.choices.first() {
+                println!("Successfully transformed text");
+                Ok(choice.text.clone())
+            } else {
+                let error = "No completion choices returned from OpenAI".to_string();
+                println!("{}", error);
+                Err(error)
+            }
+        }
+        Err(e) => {
+            let error = format!("OpenAI API error: {}", e);
+            println!("{}", error);
+            Err(error)
+        }
+    }
 }
+
 
 #[tauri::command]
 pub async fn transform_clipboard(
@@ -52,7 +90,7 @@ pub async fn transform_clipboard(
         .map_err(|e| format!("Failed to get API key: {}", e))?;
     
     // Transform the text
-    let transformed_text = crate::transform_text(&cleaned_text, prompt, &api_key).await
+    let transformed_text = transform_text(&cleaned_text, prompt, &api_key).await
         .map_err(|e| format!("Failed to transform text: {}", e))?;
 
     // Clean up transformed text before setting to clipboard
